requirements.txt
aiogram>=3.0.0rc1

python-dotenv
sqlalchemy
asyncpg
alembic